pyproject.toml
[project]
name = "research-chat"
version = "0.1.0"
description = "Routing orchestrator with deep research + text-to-SQL agents"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.6",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.0.0",
    "langchain-anthropic>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-google-genai>=3.1.0",
    "tavily-python>=0.5.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "aiosqlite>=0.20.0",
    "docker>=7.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]
